+++ modules/release.nix
@@ -85,14 +85,16 @@ in
     incremental = mkOption {
       default = false;
       type = types.bool;
-      description = "Whether to include an incremental build in otaDir";
+      description = "Whether to include an incremental build in `otaDir` output";
     };
 
     retrofit = mkOption {
       default = false;
       type = types.bool;
-      description = "Generate a retrofit OTA for upgrading a device without dynamic partitions";
-      # https://source.android.com/devices/tech/ota/dynamic_partitions/ab_legacy#generating-update-packages
+      description = ''
+        Generate a retrofit OTA for upgrading a device without dynamic partitions.
+        See also https://source.android.com/devices/tech/ota/dynamic_partitions/ab_legacy#generating-update-packages
+      '';
     };
 
     otaArgs = mkOption {
@@ -143,8 +145,9 @@ in
     );
 
     # TODO: Do this in a temporary directory. It's ugly to make build dir and ./tmp/* dir gets cleared in these scripts too.
-    # Maybe just remove this script? It's definitely complicated--and often untested
-    releaseScript = pkgs.writeScript "release.sh" (''
+    releaseScript =
+      (if (!config.signing.enable) then warn "releaseScript should be used only if signing.enable = true; Otherwise, the build might be using incorrect keys / certificate metadata" else id)
+      pkgs.writeScript "release.sh" (''
       #!${pkgs.runtimeShell}
       set -euo pipefail
 
@@ -154,12 +157,8 @@ in
         PREV_BUILDNUMBER=""
       fi
       '' + (wrapScript { keysDir="$1"; commands=''
-      if [[ "$KEYSDIR" ]]; then
-        echo Signing target files
-        ${signedTargetFilesScript { targetFiles=unsignedTargetFiles; out=signedTargetFiles.name; }}
-      else
-        echo No KEYSDIR specified. Skipping signing target files.
-      fi
+      echo Signing target files
+      ${signedTargetFilesScript { targetFiles=unsignedTargetFiles; out=signedTargetFiles.name; }}
       echo Building OTA zip
       ${otaScript { targetFiles=signedTargetFiles.name; out=ota.name; }}
       if [[ ! -z "$PREV_BUILDNUMBER" ]]; then