+++ modules/signing.nix
@@ -43,12 +43,13 @@ in
         mode = mkOption {
           type = types.strMatching "(verity_only|vbmeta_simple|vbmeta_chained|vbmeta_chained_v2)";
           default  = "vbmeta_chained";
+          description = "Mode of AVB signing to use.";
         };
 
         fingerprint = mkOption {
           type = types.strMatching "[0-9A-F]{64}";
           apply = toUpper;
-          description = "SHA256 hash of avb_pkmd.bin";
+          description = "SHA256 hash of `avb_pkmd.bin`. Should be set automatically based on file under `keyStorePath` if `signing.enable = true`";
         };
 
         verityCert = mkOption {
@@ -58,7 +59,7 @@ in
       };
 
       apex = {
-        enable = mkEnableOption "APEX signing";
+        enable = mkEnableOption "signing APEX packages";
 
         packageNames = mkOption {
           default = [];
@@ -70,7 +71,11 @@ in
 
     keyStorePath = mkOption {
       type = types.str;
-      description = "Absolute path to generated keys for signing";
+      description = ''
+        String containing absolute path to generated keys for signing.
+        This must be a _string_ and not a "nix path" to ensure that your secret keys are not imported into the public `/nix/store`.
+      '';
+      example = "/var/secrets/android-keys";
     };
   };